Add demo of compound document
Combine compound document and sparse fieldsets
functionality 

This demo fetches details about a single trip,
and the driver details can be included to skip
having to make a second API request for the driver
details.

Additionally, the fields on the driver can be limited.
Currently the serialize has 2 fields, display_name
and average_rating. There are controller tests
that demonstrate requesting all the fields for the
driver, and only the average_rating for the driver.

The driver association was set up on the trip
serializer, and a driver serializer was added.

A new route was added for a single trip, although
compound document support could be added to the
“my trips” API. This single trip API duplicates
the #show endpoint, but that was set up for demonstrating
http caching, so I left it.


https://github.com/Netflix/fast_jsonapi#compound-document

https://jsonapi.org/format/#document-compound-documents