Implement exclusion constraint reservation example in spec
Implement exclusion constraint check in SQL

The exclusion constraint prevents overlapping reservations at the same
time for the same vehicle. It is implemented using a gist index that
compares a candidate row at insert time.

On the rails side, set up a spec to show the exception is raised and
make sure the information in the exception shows the exclusion
constraint type.

Switch from Ruby schema format to SQL format for more precision. The app
should have the exact int and timestamp column types because the
constraint is relying on comparing exact times.