Partition table conversion
Using pgslice, perform a conversion of the trip_positions table to use
monthly partitions, with a partition key on the `created_at` timestamp
column

Having generated the data prior to partitioning, using pgslice fill,
copied the data into the appropriate monthly partitions. The former
unpartitioned version of the table is now called trip_positions_retired
and could be archived.