Automate partition creation
For example, every 1 month, create 6 months worth of partitions. This
will be run locally, and is mostly for demonstration purposes. The idea
is that each month it runs, since the command is idempotent, it will
append any new partitions needed when they don't exist, based on the
month it is executing in.