Fix the vehicle reservation canceled boolean
The problem was the boolean value had a default of true, which is
incorrect. The previous migration must not have worked as intended

This new migration fixes it, the column DEFAULT should be false for new
vehicle reservations. If a reservation is canceled, it would be changed
from false to true.

Also modified the fixture, and changed some of the formatting of the
dates.