Remove unneeded gems
Add necessary gems

Get the basics working for serving assets in order to serve pghero
correctly at the /pghero path

Remove some stale items from the past when the application was a
full-stack app related to assets. Remove some configuration from Blazer
since that was removed.

Compare the configuration to a newly generated rails 7 app and use the
rails 7 defaults in config/application.rb

Remove Spring configuration

Remove listen and live reload configuration that's not used

Add some notes on assets to the readme

Remove a webpack bin that's not needed