Make directory for dump and load scripts
Add README to that directory

Version control the single main data loader script mentioned in the
book. Make sure this can be re-run at any time. May need to move to a
versioned file for upcoming schema changes.

Changes to db/structure.sql. There is a "temp" schema mentioned in the
book in exercises so that snuck in to this file. It's ok either way if
it's part of the db/structure.sql dump.