Update migration
The previously generated fake data that was uploaded as the data dump,
which can be downloaded, includes data that violates this new check
constraint that was added.

Modify the check constraint to not validate existing rows. Ideally the
data dump will be updated, but it can also serve for the time being, as
an example of invalid data and how this affects adding a new constraint.