Create app_readonly role
Use the Crunchy Data post from Jonathan Katz that shows how to create a
read-only role, and how to grant the pg_read_all_data role instead of
using default privileges (due to the limitation described in the post).

Set this up for Rideshare

Confirm while connected as app_readonly that the role cannot create
tables, or update table rows.

Generate a password for the role and add it to the sample .pgpass file.

Make it a LOGIN role.