Update timeouts based on feedback
@danielabar asked about `timeout` and this prompted exploring the rails
source code. This seems to only be for sqlite, so maybe it was leftover
from originally generating this config file for sqlite. PostgreSQL and
MySQL have different parameters they support.

This is all about PostgreSQL, I added the params related to PostgreSQL
and Rails Guide and PostgreSQL documentation links.

The things we're protecting:

- Max allowed time to check out a connection from the AR pool
- Max time to wait before giving up, when all connections are busy

On PostgreSQL side we can protect against:

- Statements that would otherwise run forever in error, they should be
  canceled
- Statements that open a transaction and then are idle for too long